PWA-852: Moved out action listener into it's own class to be used as dedicated functionality. Improved building of the country list to avoid unnecessary memory consumption
